refactor: move safeCompare util from login handler to utils module
