+++ src/lib/utils.ts
@@ -1,3 +1,4 @@
+import { timingSafeEqual } from 'node:crypto'
 import type { ChatMessage, Message } from './types'
 
 export const parseSummaryToHumanReadable = (rawOutput: string): string => {
@@ -39,4 +40,13 @@ export const extractReplies = (rawOutput: string): string[] => {
         })
         .filter(Boolean)
     return replies
+}
+
+export const safeCompare = (a: string, b: string): boolean => {
+    const len = Math.max(a.length, b.length)
+    const aBuf = Buffer.alloc(len, 0)
+    const bBuf = Buffer.alloc(len, 0)
+    Buffer.from(a).copy(aBuf)
+    Buffer.from(b).copy(bBuf)
+    return timingSafeEqual(aBuf, bBuf) && a.length === b.length
 }
